@@ -23,11 +23,8 @@ packages = [
 ]
 
 [tool.poetry.dependencies]
-python = ">=3.9,<4.0"
-importlib-metadata = {version = "^4.0", python = "<3.9"}
-ansys-fluent-core = "~=0.21.dev1"
-h5py = ">=3.7.0"
-requests = "==2.31.0"
+python = ">=3.10,<4.0"
+ansys-fluent-core = "~=0.27.dev1"
 
 [tool.poetry.urls]
 "Documentation" = "https://parametric.fluent.docs.pyansys.com/"
