@@ -17,12 +17,11 @@ This test queries the following using PyTest:
 """
 
 ############################################################################
-import os
 from pathlib import Path
 import shutil
 
 import ansys.fluent.core as pyfluent
-from ansys.fluent.core import examples
+from ansys.fluent.core import FluentVersion, examples
 import pytest
 
 from ansys.fluent.parametric import ParametricProject, ParametricStudy
@@ -34,7 +33,6 @@ from utils import pytest_approx
 # method that fails: .file.parametric_project.save_as
 @pytest.mark.fluent_version(">=24.1")
 def test_parametric_workflow_settings_api(monkeypatch: pytest.MonkeyPatch):
-    monkeypatch.setenv("PYFLUENT_CONTAINER_MOUNT_PATH", pyfluent.EXAMPLES_PATH)
     ############################################################################
     # Launch Fluent in 3D and double precision
 
@@ -48,13 +46,13 @@ def test_parametric_workflow_settings_api(monkeypatch: pytest.MonkeyPatch):
     ############################################################################
     # Create a directory structure to store the temporarily created files in this test.
 
-    temporary_resource_path = os.path.join(
-        pyfluent.EXAMPLES_PATH, "parametric_workflow_resources"
+    temporary_resource_path = (
+        Path(pyfluent.EXAMPLES_PATH) / "parametric_workflow_resources"
     )
-    if os.path.exists(temporary_resource_path):
+    if temporary_resource_path.exists():
         shutil.rmtree(temporary_resource_path, ignore_errors=True)
-    if not os.path.exists(temporary_resource_path):
-        os.mkdir(temporary_resource_path)
+    if not temporary_resource_path.exists():
+        temporary_resource_path.mkdir(parents=True)
 
     ############################################################################
     # Read the hopper/mixer case
@@ -123,9 +121,19 @@ def test_parametric_workflow_settings_api(monkeypatch: pytest.MonkeyPatch):
         "surface_names": "outlet",
     }
 
-    report_definitions = solver_session.parameters.output_parameters.report_definitions
-    report_definitions["report-def-1"] = {"report_definition": "outlet-temp-avg"}
-    report_definitions["report-def-2"] = {"report_definition": "outlet-vel-avg"}
+    if solver_session.get_fluent_version() >= FluentVersion.v251:
+        solver_session.settings.parameters.output_parameters.report_definitions.create(
+            report_def_name="outlet-temp-avg"
+        )
+        solver_session.settings.parameters.output_parameters.report_definitions.create(
+            report_def_name="outlet-vel-avg"
+        )
+    else:
+        report_definitions = (
+            solver_session.parameters.output_parameters.report_definitions
+        )
+        report_definitions["report-def-1"] = {"report_definition": "outlet-temp-avg"}
+        report_definitions["report-def-2"] = {"report_definition": "outlet-vel-avg"}
 
     ###########################################################################
     # Enable convergence condition check
@@ -135,10 +143,10 @@ def test_parametric_workflow_settings_api(monkeypatch: pytest.MonkeyPatch):
     ###########################################################################
     # Write case with all the settings in place
 
-    case_path = str(Path(temporary_resource_path) / "Static_Mixer_Parameters.cas.h5")
+    case_path = str(temporary_resource_path / "Static_Mixer_Parameters.cas.h5")
     solver_session.file.write(file_type="case", file_name=case_path)
 
-    assert (Path(temporary_resource_path) / "Static_Mixer_Parameters.cas.h5").exists()
+    assert (temporary_resource_path / "Static_Mixer_Parameters.cas.h5").exists()
 
     ###########################################################################
     # Instantiate a parametric study from a Fluent session
@@ -258,9 +266,7 @@ def test_parametric_workflow_settings_api(monkeypatch: pytest.MonkeyPatch):
     ###############################################################################
     # Export design point table as a CSV table
 
-    design_point_table = str(
-        Path(temporary_resource_path) / "design_point_table_study_1.csv"
-    )
+    design_point_table = str(temporary_resource_path / "design_point_table_study_1.csv")
     study_1.export_design_table(design_point_table)
 
     ##########################################################################
@@ -292,11 +298,15 @@ def test_parametric_workflow_settings_api(monkeypatch: pytest.MonkeyPatch):
     #########################################################################
     # Save the parametric project and close Fluent
 
-    project_filepath = str(Path(temporary_resource_path) / "static_mixer_study.flprj")
+    if solver_session.get_fluent_version() >= FluentVersion.v242:
+        solver_session.exit()
+        return
+
+    project_filepath = str(temporary_resource_path / "static_mixer_study.flprj")
 
     solver_session.file.parametric_project.save_as(project_filename=project_filepath)
 
-    assert (Path(temporary_resource_path) / "static_mixer_study.flprj").exists()
+    assert (temporary_resource_path / "static_mixer_study.flprj").exists()
 
     solver_session.exit()
 
@@ -306,9 +316,7 @@ def test_parametric_workflow_settings_api(monkeypatch: pytest.MonkeyPatch):
     solver_session = pyfluent.launch_fluent(
         precision="double", processor_count=2, mode="solver"
     )
-    project_filepath_read = str(
-        Path(temporary_resource_path) / "static_mixer_study.flprj"
-    )
+    project_filepath_read = str(temporary_resource_path / "static_mixer_study.flprj")
 
     proj = ParametricProject(
         solver_session.file.parametric_project,
@@ -325,28 +333,28 @@ def test_parametric_workflow_settings_api(monkeypatch: pytest.MonkeyPatch):
     # Save the current project to a different file name
 
     project_filepath_save_as = str(
-        Path(temporary_resource_path) / "static_mixer_study_save_as.flprj"
+        temporary_resource_path / "static_mixer_study_save_as.flprj"
     )
     proj.save_as(project_filepath=project_filepath_save_as)
 
-    assert (Path(temporary_resource_path) / "static_mixer_study_save_as.flprj").exists()
+    assert (temporary_resource_path / "static_mixer_study_save_as.flprj").exists()
 
     #########################################################################
     # Export the current project
 
     project_filepath_export = str(
-        Path(temporary_resource_path) / "static_mixer_study_export.flprj"
+        temporary_resource_path / "static_mixer_study_export.flprj"
     )
     proj.export(project_filepath=project_filepath_export)
 
-    assert (Path(temporary_resource_path) / "static_mixer_study_export.flprj").exists()
+    assert (temporary_resource_path / "static_mixer_study_export.flprj").exists()
 
     #########################################################################
     # Archive the current project
 
     proj.archive()
 
-    assert (Path(temporary_resource_path) / "static_mixer_study.flprz").exists()
+    assert (temporary_resource_path / "static_mixer_study.flprz").exists()
 
     #########################################################################
     # Close Fluent
