@@ -1,4 +1,3 @@
-import ansys.fluent.core as pyfluent
 from ansys.fluent.core import examples
 import pytest
 
@@ -6,13 +5,12 @@ from ansys.fluent.parametric import ParametricSession
 
 
 def test_parametric_session(monkeypatch: pytest.MonkeyPatch):
-    monkeypatch.setenv("PYFLUENT_CONTAINER_MOUNT_PATH", pyfluent.EXAMPLES_PATH)
     project_filepath = examples.download_file(
         "project-elbow-param.flprz",
         "pyfluent/mixing_elbow",
         return_without_path=False,
     )
-    session = ParametricSession(project_filepath=project_filepath, initialize=False)
+    session = ParametricSession(project_filepath=project_filepath)
     assert "Static_Mixer_main-Solve" in session.studies
     design_points = session.studies["Static_Mixer_main-Solve"].design_points
     assert "Base DP" in design_points
